addon/-private renames
