+++ addon/-private/lifecycle/model/index.js
