+++ addon/-private/lifecycle/models/property.js
@@ -7,7 +7,7 @@ const reusable = () => true;
 const create = opts => function(key) {
   let parent = this;
   let owner = getOwner(this);
-  return owner.factoryFor('zuglet:less-experimental/models/internal').create({ parent, key, opts });
+  return owner.factoryFor('zuglet:lifecycle/models/internal').create({ parent, key, opts });
 }
 
 export default opts => {
