+++ addon/-private/lifecycle/observed/index.js
