+++ addon/-private/lifecycle/observed/internal/dynamic.js
