+++ addon/-private/lifecycle/observed/observer-for.js
