+++ addon/-private/lifecycle/observed/resolve-observers.js
