+++ addon/-private/lifecycle/route/hooks.js
