+++ addon/-private/lifecycle/route/index.js
