+++ addon/-private/lifecycle/util/inline.js
