+++ addon/-private/lifecycle/util/object-observer.js
